Split the main class logic into several helper classes
